Revert "an attempt to add ICA routers, but hitting limits"
This reverts commit 7a49aa0b70b979eee19df21e553e950e6c8c6f0b.
