+++ typescript/infra/config/environments/mainnet3/agent.ts
@@ -22,14 +22,13 @@ import { getDomainId } from '../../registry.js';
 
 import { environment } from './chains.js';
 import { helloWorld } from './helloworld.js';
-import aaveSenderAddresses from './misc-artifacts/aave-sender-addresses.json' assert { type: 'json' };
-import everclearSenderAddresses from './misc-artifacts/everclear-sender-addresses.json' assert { type: 'json' };
-import { getInterchainAccountRouters } from './misc-artifacts/ica.js';
-import merklyEthAddresses from './misc-artifacts/merkly-eth-addresses.json' assert { type: 'json' };
-import merklyNftAddresses from './misc-artifacts/merkly-eth-addresses.json' assert { type: 'json' };
-import merklyErc20Addresses from './misc-artifacts/merkly-eth-addresses.json' assert { type: 'json' };
-import veloMessageModuleAddresses from './misc-artifacts/velo-message-module-addresses.json' assert { type: 'json' };
-import veloTokenBridgeAddresses from './misc-artifacts/velo-token-bridge-addresses.json' assert { type: 'json' };
+import aaveSenderAddresses from './misc-artifacts/aave-sender-addresses.json';
+import everclearSenderAddresses from './misc-artifacts/everclear-sender-addresses.json';
+import merklyEthAddresses from './misc-artifacts/merkly-eth-addresses.json';
+import merklyNftAddresses from './misc-artifacts/merkly-eth-addresses.json';
+import merklyErc20Addresses from './misc-artifacts/merkly-eth-addresses.json';
+import veloMessageModuleAddresses from './misc-artifacts/velo-message-module-addresses.json';
+import veloTokenBridgeAddresses from './misc-artifacts/velo-token-bridge-addresses.json';
 import {
   mainnet3SupportedChainNames,
   supportedChainNames,
@@ -443,10 +442,6 @@ const metricAppContextsGetter = (): MetricAppContext[] => {
       ),
     },
     {
-      name: 'ica_router',
-      matchingList: routerMatchingList(getInterchainAccountRouters()),
-    },
-    {
       name: 'merkly_erc20',
       matchingList: routerMatchingList(merklyErc20Addresses),
     },
@@ -487,21 +482,21 @@ const metricAppContextsGetter = (): MetricAppContext[] => {
 const relayerResources = {
   requests: {
     cpu: '14000m',
-    memory: '20G',
+    memory: '15Gi',
   },
 };
 
 const validatorResources = {
   requests: {
     cpu: '500m',
-    memory: '1G',
+    memory: '1Gi',
   },
 };
 
 const scraperResources = {
   requests: {
     cpu: '2000m',
-    memory: '4G',
+    memory: '4Gi',
   },
 };
 
@@ -627,7 +622,7 @@ const neutron: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '7c0c967-20241218-173053',
+      tag: '25a927d-20241114-171323',
     },
     blacklist,
     gasPaymentEnforcement,
