+++ typescript/infra/src/utils/utils.ts
@@ -196,12 +196,6 @@ export function readYaml<T>(filepath: string): T {
   return yamlParse(readFileAtPath(filepath)) as T;
 }
 
-export function removeFile(filepath: string) {
-  if (fs.existsSync(filepath)) {
-    fs.unlinkSync(filepath);
-  }
-}
-
 export function assertRole(roleStr: string) {
   const role = roleStr as Role;
   if (!Object.values(Role).includes(role)) {
